[IMP] budget_appropriation_summary: migrate dashboard from budget_appropriation_report
