+++ budget_appropriation_report/__manifest__.py
@@ -8,15 +8,8 @@
     "website": "https://github.com/aginix/kmitl",
     "depends": ["budget_appropriation", "web"],
     'data': [
-        "data/budget_appropriation_report_action.xml",
         "data/paper_format.xml",
-        "views/menus.xml",
     ],
-    "assets": {
-        "web.assets_backend": [
-            "budget_appropriation_report/static/src/**/*",
-        ],
-    },
     "auto_install": False,
     "application": False,
     "installable": True,
