+++ budget_appropriation_summary/static/src/components/budget_appropriation_dashboard/budget_appropriation_dashboard.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <templates xml:space="preserve">
-    <t t-name="budget_appropriation_report.BudgetAppropriationDashboard" owl="1">
+    <t t-name="budget_appropriation_summary.BudgetAppropriationDashboard" owl="1">
         <div class="o_action budget_appropriation_dashboard">
             <ControlPanel display="controlPanelDisplay">
                 <t t-set-slot="control-panel-top-right">
