+++ budget_appropriation_summary/views/menu.xml
@@ -9,6 +9,14 @@
     />
 
     <menuitem
+        id="menu_budget_appropriation_dashboard"
+        name="Dashboard การจัดสรรงบประมาณ"
+        parent="menu_budget_appropriation_summary_root"
+        action="action_budget_appropriation_dashboard"
+        sequence="1"
+    />
+
+    <menuitem
         id="menu_budget_appropriation_compilation"
         name="รวมเล่มหน่วยงาน"
         parent="menu_budget_appropriation_summary_root"
